Add FastDDS example JSON file with binary paths and dependencies.
This commit adds a new JSON file, `FastDDS_Examples.json`, that contains information about various FastDDS examples. Each example includes the binary path and size, as well as a list of dependencies with their respective sizes. This file will be useful for tracking and managing the examples in the FastDDS project.